Feat new application

+++ backend/src/main/java/net/tamasnovak/repositories/AccountRepository.java
@@ -1,10 +1,10 @@
 package net.tamasnovak.repositories;
 
-import net.tamasnovak.entities.Account;
+import net.tamasnovak.entities.account.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AccountRepository extends JpaRepository<Account, Integer> {
+public interface AccountRepository extends JpaRepository<Account, Long> {
   Optional<Account> findByEmail(String email);
 }
